+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Changed
 - add `useMaskLayerVisibility` custom hook
 - add `getChangeCompareLayerRasterFunction` helper function
+- add `useDataFromSelectedImageryScene` custom hook
 
 ### Fixed